requirements.txt
Django>=5.0
djangorestframework>=3.15.0
django-redis>=5.4.0
django-cors-headers>=4.5.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.0
django-phonenumber-field>=8.0.0
phonenumbers>=9.0.0
celery>=5.3.0
twilio>=9.0.0
sendgrid>=6.10.0